Update index.html
At various times in the last decade and change, I've been a homebrewer and a huge craft beer aficionado. Over the last two years, I've really slowed down and started to appreciate local spirits more. I'm a big fan of what Family Jones, The Block Distilling, and Ironton are up to in Denver.